refactor: rename process_resource to analyse_resource
