+++ udata_hydra/db/check.py
@@ -1,5 +1,7 @@
 from typing import Optional
 
+import asyncpg
+
 from udata_hydra import context
 from udata_hydra.db import (
     compute_insert_query,
@@ -12,7 +14,7 @@ class Check:
     """Represents a check in the "checks" DB table"""
 
     @classmethod
-    async def get_by_id(cls, check_id: int, with_deleted: bool = False) -> Optional[dict]:
+    async def get_by_id(cls, check_id: int, with_deleted: bool = False) -> Optional[asyncpg.Record]:
         pool = await context.pool()
         async with pool.acquire() as connection:
             q = """
@@ -27,7 +29,7 @@ class Check:
     @classmethod
     async def get_by_resource_id(
         cls, resource_id: str, with_deleted: bool = False
-    ) -> Optional[dict]:
+    ) -> Optional[asyncpg.Record]:
         pool = await context.pool()
         async with pool.acquire() as connection:
             q = """
@@ -42,7 +44,7 @@ class Check:
     @classmethod
     async def get_latest(
         cls, url: Optional[str] = None, resource_id: Optional[str] = None
-    ) -> Optional[dict]:
+    ) -> Optional[asyncpg.Record]:
         column: str = "url" if url else "resource_id"
         pool = await context.pool()
         async with pool.acquire() as connection:
