+++ udata_hydra/routes/checks.py
@@ -2,6 +2,7 @@ import json
 from typing import Optional
 
 import aiohttp
+import asyncpg
 from aiohttp import web
 from marshmallow import ValidationError
 
@@ -16,7 +17,7 @@ from udata_hydra.utils import get_request_params
 async def get_latest_check(request: web.Request) -> web.Response:
     """Get the latest check for a given URL or resource_id"""
     url, resource_id = get_request_params(request, params_names=["url", "resource_id"])
-    data: Optional[dict] = await Check.get_latest(url, resource_id)
+    data: Optional[asyncpg.Record] = await Check.get_latest(url, resource_id)
     if not data:
         raise web.HTTPNotFound()
     if data["deleted"]:
@@ -46,7 +47,7 @@ async def create_check(request: web.Request) -> web.Response:
 
     # Get URL from resource_id
     try:
-        resource: dict = await Resource.get(resource_id, "url")
+        resource = await Resource.get(resource_id, "url")
         url: str = resource["url"]
     except Exception:
         raise web.HTTPNotFound(text=f"Couldn't find URL for resource {resource_id}")
@@ -61,7 +62,7 @@ async def create_check(request: web.Request) -> web.Response:
         )
         context.monitor().refresh(status)
 
-    check: Optional[dict] = await Check.get_latest(url, resource_id)
+    check: Optional[asyncpg.Record] = await Check.get_latest(url, resource_id)
     if not check:
         raise web.HTTPBadRequest(text=f"Check not created, status: {status}")
